# src/fleet/ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/tasks": {
      "get": {
        "summary": "List Tasks",
        "description": "List task summaries, active first then recently-closed (FR-07).",
        "operationId": "list_tasks_api_tasks_get",
        "parameters": [
          {
            "name": "closed_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 300,
              "title": "Closed Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Task",
        "description": "Create a task via the queue; 201 with the new id.",
        "operationId": "create_task_api_tasks_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "description": "One task summary, overlaid with beads status/priority/depends_on.",
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Task",
        "description": "Delete a task via the queue.",
        "operationId": "delete_task_api_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/{task_id}/logs": {
      "get": {
        "summary": "Get Task Logs",
        "description": "Parsed log.jsonl lines, optionally filtered by level (FR-17).",
        "operationId": "get_task_logs_api_tasks__task_id__logs_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/stderr": {
      "get": {
        "summary": "Get Task Stderr",
        "description": "Raw stderr content, empty when absent (FR-18).",
        "operationId": "get_task_stderr_api_tasks__task_id__stderr_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/files": {
      "get": {
        "summary": "Get Task Files",
        "description": "Per-file read/edit/write counts from the event scan (FR-20).",
        "operationId": "get_task_files_api_tasks__task_id__files_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/events": {
      "get": {
        "summary": "Get Task Events",
        "description": "Whole-task event history across attempts, paged (tail by default).",
        "operationId": "get_task_events_api_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ContentResponse",
        "description": "Single-text envelope (attempt summary/prompt/log/state, stderr)."
      },
      "CreateTaskResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CreateTaskResponse",
        "description": "New task id for POST /api/tasks."
      },
...
      "FileListResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileOp"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "FileListResponse",
        "description": "Envelope for GET /api/tasks/{id}/files."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobArtifacts": {
        "properties": {
          "research": {
            "type": "boolean",
            "title": "Research"
          },
          "design": {
            "type": "boolean",
            "title": "Design"
          },
          "tasks": {
            "type": "boolean",
            "title": "Tasks"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved"
          }
        },
        "type": "object",
        "required": [
          "research",
          "design",
          "tasks",
          "approved"
        ],
        "title": "JobArtifacts",
        "description": "Presence flags for the job worker's RESEARCH/DESIGN/tasks/APPROVED."
      },
...
      "LogListResponse": {
        "properties": {
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LogLine"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "lines"
        ],
        "title": "LogListResponse",
        "description": "Envelope for GET /api/tasks/{id}/logs."
      },
      "OkResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "description": "Generic mutation acknowledgement ({\"ok\": true})."
      },
...
      "RunStep": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "fail",
              "outcome"
            ],
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "name",
          "started_at",
          "ended_at",
          "status",
          "reason"
        ],
        "title": "RunStep",
        "description": "One entry of the current attempt's run.json steps."
      },
...
      "TaskAttempt": {
        "properties": {
          "n": {
            "type": "integer",
            "title": "N"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "coder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coder"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outcome"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "peak_context_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Context Pct"
          },
          "context_badge": {
            "type": "boolean",
            "title": "Context Badge"
          },
          "files_touched": {
            "type": "integer",
            "title": "Files Touched"
          },
          "commits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commits"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "has_summary": {
            "type": "boolean",
            "title": "Has Summary"
          },
          "has_prompt": {
            "type": "boolean",
            "title": "Has Prompt"
          }
        },
        "type": "object",
        "required": [
          "n",
          "kind",
          "mode",
          "coder",
          "model",
          "started_at",
          "ended_at",
          "duration_sec",
          "outcome",
          "reason",
          "peak_context_pct",
          "context_badge",
          "files_touched",
          "commits",
          "result",
          "has_summary",
          "has_prompt"
        ],
        "title": "TaskAttempt",
        "description": "One row of the attempts timeline."
      },
...
      "TaskDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "coder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coder"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "elapsed_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Sec"
          },
          "idle_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idle Sec"
          },
          "events": {
            "type": "integer",
            "title": "Events"
          },
          "context_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Tokens"
          },
          "context_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Pct"
          },
          "context_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Limit"
          },
          "last_event_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Kind"
          },
          "last_event_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Detail"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "blocked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked At"
          },
          "ignore_until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ignore Until"
          },
          "ignored": {
            "type": "boolean",
            "title": "Ignored"
          },
          "rounds": {
            "$ref": "#/components/schemas/TaskRounds"
          },
          "restarts": {
            "type": "integer",
            "title": "Restarts"
          },
          "context_rounds": {
            "type": "integer",
            "title": "Context Rounds"
          },
          "compactions": {
            "type": "integer",
            "title": "Compactions"
          },
          "peak_context_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak Context Pct"
          },
          "last_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Outcome"
          },
          "last_outcome_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Outcome Reason"
          },
          "last_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "state_excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State Excerpt"
          },
          "worker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker"
          },
          "job_phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Phase"
          },
          "job_artifacts": {
            "$ref": "#/components/schemas/JobArtifacts"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/RunStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "lease": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskLease"
              },
              {
                "type": "null"
              }
            ]
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/TaskAttempt"
            },
            "type": "array",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "status",
          "cwd",
          "coder",
          "model",
          "priority",
          "depends_on",
          "created_at",
          "started_at",
          "ended_at",
          "elapsed_sec",
          "idle_sec",
          "events",
          "context_tokens",
          "context_pct",
          "context_limit",
          "last_event_kind",
          "last_event_detail",
          "blocked_reason",
          "blocked_at",
          "ignore_until",
          "ignored",
          "rounds",
          "restarts",
          "context_rounds",
          "compactions",
          "peak_context_pct",
          "last_outcome",
          "last_outcome_reason",
          "last_action",
          "result",
          "state_excerpt",
          "worker",
          "job_phase",
          "job_artifacts",
          "steps",
          "lease",
          "attempts"
        ],
        "title": "TaskDetail",
        "description": "One task with its attempts timeline (GET /api/tasks/{id})."
      },
      "TaskEventsResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/StreamEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "total",
          "offset",
          "events"
        ],
        "title": "TaskEventsResponse",
        "description": "Paged envelope for GET /api/tasks/{id}/events."
      },
      "TaskLease": {
        "properties": {
          "heartbeat_at": {
            "type": "string",
            "title": "Heartbeat At"
          },
          "lease_until": {
            "type": "string",
            "title": "Lease Until"
          },
          "alive": {
            "type": "boolean",
            "title": "Alive"
          }
        },
        "type": "object",
        "required": [
          "heartbeat_at",
          "lease_until",
          "alive"
        ],
        "title": "TaskLease",
        "description": "Claim lease from the latest attempt's run.json, null when absent."
      },
      "TaskListResponse": {
        "properties": {
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskDetail"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "tasks"
        ],
        "title": "TaskListResponse",
        "description": "Envelope for GET /api/tasks."
      },
      "TaskResult": {
        "properties": {
          "schema": {
            "type": "integer",
            "title": "Schema"
          },
          "status": {
            "type": "string",
            "enum": [
              "done",
              "partial",
              "blocked"
            ],
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "commits": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commits"
          },
          "tests": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tests"
          },
          "open_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open Questions"
          },
          "next_step": {
            "type": "string",
            "title": "Next Step"
          },
          "blocked_reason": {
            "type": "string",
            "title": "Blocked Reason"
          },
          "followups": {
            "items": {
              "$ref": "#/components/schemas/FollowupItem"
            },
            "type": "array",
            "title": "Followups"
          }
        },
        "type": "object",
        "required": [
          "schema",
          "status",
          "summary",
          "commits",
          "tests",
          "open_questions",
          "next_step",
          "blocked_reason"
        ],
        "title": "TaskResult",
        "description": "The worker's declared outcome (live RESULT.json or latest snapshot)."
      },
      "TaskRounds": {
        "properties": {
          "failure": {
            "type": "integer",
            "title": "Failure"
          },
          "stall": {
            "type": "integer",
            "title": "Stall"
          },
          "context": {
            "type": "integer",
            "title": "Context"
          },
          "partial": {
            "type": "integer",
            "title": "Partial"
          },
          "noclose": {
            "type": "integer",
            "title": "Noclose"
          }
        },
        "type": "object",
        "required": [
          "failure",
          "stall",
          "context",
          "partial",
          "noclose"
        ],
        "title": "TaskRounds",
        "description": "Trailing-streak retry rounds, from attempts.jsonl."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}